Update os.cpp
Progress so far. Difficult to implement Svc without knowing how scheduler is sending the service requests. But I have a basic outline of how each service will go about its operation. Need to figure out how to access the jobs; It is difficult because nothing is running; the jobs are just in memory and I don't know what they want.